refactor: Move MenuMapper to /global/mapper (#23)
